Added validations, constants and junits

+++ src/main/java/com/silverbars/application/beans/OrderBean.java
@@ -1,5 +1,7 @@
 package com.silverbars.application.beans;
 
+import java.math.BigDecimal;
+
 /**
  * Model class for order
  * 
@@ -11,7 +13,7 @@ public class OrderBean {
 	private Integer id;
 	private String userId;
 	private Double quantityKg;
-	private Double pricePerKg;
+	private BigDecimal pricePerKg;
 	private OrderType orderType;
 
 	public Integer getId() {
@@ -38,11 +40,11 @@ public class OrderBean {
 		this.quantityKg = quantityKg;
 	}
 
-	public Double getPricePerKg() {
+	public BigDecimal getPricePerKg() {
 		return pricePerKg;
 	}
 
-	public void setPricePerKg(Double pricePerKg) {
+	public void setPricePerKg(BigDecimal pricePerKg) {
 		this.pricePerKg = pricePerKg;
 	}
